test(Slider): update test cases
